requirements.txt
tencentcloud-sdk-python-asr
tencentcloud-sdk-python-common
pvporcupine
pvrecorder
pydub
wave
numpy
openai
edge-tts